BookApp/Services/BookService.cs
using BookApp.DAL;
using BookApp.Models;

namespace BookApp.Services
{
    public class BookService : IBookService
    {
        private IBookRepository _bookRepository;

        public BookService(IBookRepository bookRepository)
        {
            _bookRepository = bookRepository;
        }

        public void CreateBook(Book book)
        {
            _bookRepository.CreateBook(book);
        }

        public void DeleteBook(int id)
        {
            _bookRepository.DeleteBook(id);
        }

        public Book GetBookByID(int id)
        {
            return _bookRepository.GetBookByID(id);
        }

        public IEnumerable<Book> GetBooks()
        {
            return _bookRepository.GetBooks();
        }

        public void UpdateBook(Book book)
        {
            _bookRepository.UpdateBook(book);
        }

        public void Save()
        {
            _bookRepository.Save();
        }
    }
}